read segmentation value from current zoomstep in mappings view; enable copy-segmentation-id-shortcut for non-volume-tracings
