+++ app/assets/javascripts/oxalis/controller/combinations/volumetracing_plane_controller.js
@@ -14,9 +14,8 @@ import {
   calculateGlobalPos,
 } from "oxalis/controller/viewmodes/plane_controller";
 import Model from "oxalis/model";
-import { getPosition } from "oxalis/model/accessors/flycam_accessor";
+import { getPosition, getRequestLogZoomStep } from "oxalis/model/accessors/flycam_accessor";
 import { setPositionAction } from "oxalis/model/actions/flycam_actions";
-import Clipboard from "clipboard-js";
 import Toast from "libs/toast";
 import {
   createCellAction,
@@ -183,6 +182,8 @@ class VolumeTracingPlaneController extends PlaneControllerClass {
         if (event.shiftKey) {
           const cellId = Model.getSegmentationBinary().cube.getDataValue(
             this.calculateGlobalPos(pos),
+            null,
+            getRequestLogZoomStep(Store.getState()),
           );
           this.handleCellSelection(cellId);
         }
@@ -208,20 +209,6 @@ class VolumeTracingPlaneController extends PlaneControllerClass {
   getKeyboardControls(): Object {
     return _.extend(super.getKeyboardControls(), {
       c: () => Store.dispatch(createCellAction()),
-      "ctrl + i": async event => {
-        const mousePosition = Store.getState().temporaryConfiguration.mousePosition;
-        if (mousePosition) {
-          const [x, y] = mousePosition;
-          const globalMousePosition = calculateGlobalPos({ x, y });
-          const cube = Model.getSegmentationBinary().cube;
-          const mapping = event.altKey ? cube.mapping : null;
-          const hoveredId = cube.getDataValue(globalMousePosition, mapping);
-          await Clipboard.copy(String(hoveredId));
-          Toast.success(`Cell id ${hoveredId} copied to clipboard.`);
-        } else {
-          Toast.warning("No cell under cursor.");
-        }
-      },
     });
   }
